wip: move password concerns to authentication
