+++ src/kairos/authentication/infrastructure/security/password/hash-password.ts
@@ -1,3 +1,3 @@
-import { passwordHasher } from '@/kairos/shared/security/password/password-hasher';
+import { passwordHasher } from '@/kairos/authentication/infrastructure/security/password/password-hasher';
 
 export const hashPassword = (plainPassword: string): Promise<string> => passwordHasher.hash(plainPassword);
