+++ src/kairos/authentication/infrastructure/security/password/verify-password.test.ts
@@ -1,5 +1,5 @@
 import { invalidCredentialsError } from '@/kairos/authentication/domain/errors';
-import { passwordHasher } from '@/kairos/shared/security/password/password-hasher';
+import { passwordHasher } from '@/kairos/authentication/infrastructure/security/password/password-hasher';
 import { describe, expect, it } from 'vitest';
 import { createVerifyPassword } from './verify-password';
